Let Sync from Data work with archives.  RRR
This works with any archive that has only one version of a
missing or mismatched file. Since the in-place update process uses
recursion to find files, files with many versions would all be updated.
This works with BAIN simple structured 7z and zip archives.
Rar archives are not supported. Some fomod, wizard and complex BAIN
archives will work but many won't. The user will get a warning message
when attempting to Sync from Data an archive containing file with more
than one version. Also some renamed doc files are not supported,
e.g. readme.txt.

Because 7zip can't update rar archives in-place, rar archives are not supported.

Some doc files are renamed when installed so those installed filenames
won't match the one in the archive so can't be updated.

Larger archives may take a few moments to update. The progress bar does not
advance since updating is done in one call to 7z.

Note: Files that are lower in the hierarchy appear in the Mismatched tab.
These therefore appear in the Sync from Data prompt. However these
mismatched files aren't from the current archive. They are from packages
that are overriding them. Therefore Syncing will import them into the
selected archive! This is still an issue.

Assumed bug: If you Sync from Data a Project that isn't installed,
the syncing will remove all files from the Project. The could also
happen with archives. Now Sync from Data is only enabled for installed packages.

The code consistently calls this functionality Sync from Data.
However the Advanced Readme refers to it as Sync From Data.
Sync From Data is a clearer description. I've taken the liberty to
change the code to use the better name 'Sync From Data'.

=======A rejected design===========
Use relative paths on the 7z command line to sync to the archive
and don't require unique names. The problem here is the relative paths
specified on the command line must match the internal paths in the archive.
This works for simple structured archives but no others.
However the relative paths in most complex structures, fomods and wizard
don't match the internal relative paths in the archive.

An example why complex structured archives relative paths are not compatible.

This complex installer contains:
  00 Core/sound/kitten.wav
  01 Quiet/sound/kitten.wav
00 Core is the default and 01 Quiet has a kitten that doesn't purr. Note: the filenames are not unique.

When installed, the game/Data folder has:
   sound/kitten.wav

If kitten.wav is changed or deleted, on the command line you can't use
relative paths to uniquely identify which file in the archive needs updating.

The relative path on this command line works with local pathing but
doesn't match any file in the archive.
  7z u archive.7z 'sound/kitten.wav'

The relative path here matches the internal pathing but not local pathing.
  7z u archive.7z '00 Core/sound/kitten.wav'

Since this doesn't handle rar archives it only
partially closes # 391 <---- RRR

# Conflicts:
#	Mopy/bash/bosh/bain.py