Rework locale handling and translations RRR
Yup, again. This time it works on py3 (and hopefully Linux too).

Also renamed our l10n files to be less random and to use the proper .po
extension. The .txt is a leftover from Wrye Mash, which used its own
bespoke translation format. We use gettext instead, for which the
standard extension is .po. In fact, some gettext editors (e.g. poedit)
refuse to open a gettext file if it doesn't end in .po.

This also obsoletes the 'Rename to .po' checkbox, so I dropped that too.

Under # 500, # 460, # 243 <--- RRR